--- examples/legacy/scaling-criteo/docker-compose-fastai.yml
@@ -1,20 +1,0 @@
-# (1) Use this file for docker runtime configurations that are common to both
-# development and deployment.
-
-# `version : '2.3'` lets us use the `runtime=nvidia` configuration so that our
-# containers can interact with the GPU(s).
-version: '2.3'
-
-volumes:
-  models:
-
-services:
-  lab:
-    runtime: nvidia
-    image: nvcr.io/nvidia/merlin/merlin-pytorch:22.06
-    command: "/bin/bash -c 'pip install jupyterlab jupytext pydot && apt-get update && apt-get install -y tree && python -m ipykernel install --user --name=merlin && jupyter notebook --no-browser --allow-root --port=8888 --ip=0.0.0.0 --NotebookApp.token='demotoken' --NotebookApp.allow_origin='*' --notebook-dir=/'"
-    volumes:
-      - models:/models
-      - /raid/:/raid/
-    ports:
-      - 8888:8888
